初始文件创建
新建了clue、display、file、reaction等文件，将display和鼠标事件函数移出
