+++ Makefile.win
@@ -4,11 +4,11 @@
 CPP      = g++.exe -D__DEBUG__
 CC       = gcc.exe -D__DEBUG__
 WINDRES  = windres.exe
-OBJ      = build/main.o build/exceptio.o build/genlib.o build/graphics.o build/imgui.o build/linkedlist.o build/random.o build/simpio.o build/strlib.o
-LINKOBJ  = build/main.o build/exceptio.o build/genlib.o build/graphics.o build/imgui.o build/linkedlist.o build/random.o build/simpio.o build/strlib.o
+OBJ      = build/main.o build/exceptio.o build/genlib.o build/graphics.o build/imgui.o build/linkedlist.o build/random.o build/simpio.o build/strlib.o build/clue.o build/display.o build/file.o build/reaction.o
+LINKOBJ  = build/main.o build/exceptio.o build/genlib.o build/graphics.o build/imgui.o build/linkedlist.o build/random.o build/simpio.o build/strlib.o build/clue.o build/display.o build/file.o build/reaction.o
 LIBS     = -L"D:/Dev-Cpp/MinGW64/lib" -L"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib" -static-libgcc -mwindows -g3
-INCS     = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"D:/ZJUrecord/课程作业/C大程/MazeGame/include"
-CXXINCS  = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++" -I"D:/ZJUrecord/课程作业/C大程/MazeGame/include"
+INCS     = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"D:/ZJUrecord/课程作业/C大程/MazeGame_Git/MazeGame/include" -I"D:/ZJUrecord/课程作业/C大程/MazeGame_Git/MazeGame/inc"
+CXXINCS  = -I"D:/Dev-Cpp/MinGW64/include" -I"D:/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"D:/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++" -I"D:/ZJUrecord/课程作业/C大程/MazeGame_Git/MazeGame/include" -I"D:/ZJUrecord/课程作业/C大程/MazeGame_Git/MazeGame/inc"
 BIN      = bin/MazeGame.exe
 CXXFLAGS = $(CXXINCS) -g3
 CFLAGS   = $(INCS) -g3
@@ -50,3 +50,15 @@ build/simpio.o: libgraphics/simpio.c
 
 build/strlib.o: libgraphics/strlib.c
 	$(CC) -c libgraphics/strlib.c -o build/strlib.o $(CFLAGS)
+
+build/clue.o: src/clue.c
+	$(CC) -c src/clue.c -o build/clue.o $(CFLAGS)
+
+build/display.o: src/display.c
+	$(CC) -c src/display.c -o build/display.o $(CFLAGS)
+
+build/file.o: src/file.c
+	$(CC) -c src/file.c -o build/file.o $(CFLAGS)
+
+build/reaction.o: src/reaction.c
+	$(CC) -c src/reaction.c -o build/reaction.o $(CFLAGS)
