+++ MazeGame.dev
@@ -4,7 +4,7 @@ Name=MazeGame
 Type=0
 Ver=2
 ObjFiles=
-Includes=D:\ZJUrecord\课程作业\C大程\MazeGame\include
+Includes=D:\ZJUrecord\课程作业\C大程\MazeGame_Git\MazeGame\include;D:\ZJUrecord\课程作业\C大程\MazeGame_Git\MazeGame\inc
 Libs=
 PrivateResource=
 ResourceIncludes=
@@ -24,12 +24,12 @@ HostApplication=
 UseCustomMakefile=0
 CustomMakefile=
 CommandLine=
-Folders=bin,build,include,libgraphics,src
+Folders=bin,build,inc,include,libgraphics,src
 IncludeVersionInfo=0
 SupportXPThemes=0
 CompilerSet=1
 CompilerSettings=0000000000000000001000000
-UnitCount=21
+UnitCount=29
 
 [VersionInfo]
 Major=1
@@ -61,9 +61,9 @@ OverrideBuildCmd=0
 BuildCmd=
 
 [Unit21]
-FileName=include\MyHeader.h
+FileName=libgraphics\imgui.c
 CompileCpp=0
-Folder=include
+Folder=libgraphics
 Compile=1
 Link=1
 Priority=1000
@@ -180,17 +180,7 @@ Priority=1000
 OverrideBuildCmd=0
 BuildCmd=
 
-[Unit13]
-FileName=libgraphics\exceptio.c
-CompileCpp=0
-Folder=libgraphics
-Compile=1
-Link=1
-Priority=1000
-OverrideBuildCmd=0
-BuildCmd=
-
-[Unit14]
+[Unit19]
 FileName=libgraphics\genlib.c
 CompileCpp=0
 Folder=libgraphics
@@ -200,10 +190,30 @@ Priority=1000
 OverrideBuildCmd=0
 BuildCmd=
 
-[Unit15]
-FileName=libgraphics\graphics.c
+[Unit13]
+FileName=inc\clue.h
 CompileCpp=0
-Folder=libgraphics
+Folder=inc
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit14]
+FileName=inc\display.h
+CompileCpp=0
+Folder=inc
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit15]
+FileName=inc\file.h
+CompileCpp=0
+Folder=inc
 Compile=1
 Link=1
 Priority=1000
@@ -211,9 +221,9 @@ OverrideBuildCmd=0
 BuildCmd=
 
 [Unit16]
-FileName=libgraphics\imgui.c
+FileName=inc\MyHeader.h
 CompileCpp=0
-Folder=libgraphics
+Folder=inc
 Compile=1
 Link=1
 Priority=1000
@@ -221,9 +231,9 @@ OverrideBuildCmd=0
 BuildCmd=
 
 [Unit17]
-FileName=libgraphics\linkedlist.c
+FileName=inc\reaction.h
 CompileCpp=0
-Folder=libgraphics
+Folder=inc
 Compile=1
 Link=1
 Priority=1000
@@ -231,17 +241,7 @@ OverrideBuildCmd=0
 BuildCmd=
 
 [Unit18]
-FileName=libgraphics\random.c
-CompileCpp=0
-Folder=libgraphics
-Compile=1
-Link=1
-Priority=1000
-OverrideBuildCmd=0
-BuildCmd=
-
-[Unit19]
-FileName=libgraphics\simpio.c
+FileName=libgraphics\exceptio.c
 CompileCpp=0
 Folder=libgraphics
 Compile=1
@@ -251,9 +251,89 @@ OverrideBuildCmd=0
 BuildCmd=
 
 [Unit20]
+FileName=libgraphics\graphics.c
+CompileCpp=0
+Folder=libgraphics
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit22]
+FileName=libgraphics\linkedlist.c
+CompileCpp=0
+Folder=libgraphics
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit23]
+FileName=libgraphics\random.c
+CompileCpp=0
+Folder=libgraphics
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit24]
+FileName=libgraphics\simpio.c
+CompileCpp=0
+Folder=libgraphics
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit25]
 FileName=libgraphics\strlib.c
 CompileCpp=0
 Folder=libgraphics
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit26]
+FileName=src\clue.c
+CompileCpp=0
+Folder=src
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit27]
+FileName=src\display.c
+CompileCpp=0
+Folder=src
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit28]
+FileName=src\file.c
+CompileCpp=0
+Folder=src
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit29]
+FileName=src\reaction.c
+CompileCpp=0
+Folder=src
 Compile=1
 Link=1
 Priority=1000
