+++ MazeGame.layout
@@ -1,8 +1,63 @@
 [Editors]
-Order=0
-Focused=0
+Order=0,28,27,26,25,6,12,15
+Focused=28
 [Editor_0]
-CursorCol=19
-CursorRow=30
-TopLine=20
+CursorCol=15
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_20]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_16]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_17]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_14]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_12]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_6]
+CursorCol=1
+CursorRow=148
+TopLine=142
+LeftChar=1
+[Editor_15]
+CursorCol=15
+CursorRow=10
+TopLine=1
+LeftChar=1
+[Editor_25]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_26]
+CursorCol=1
+CursorRow=2
+TopLine=1
+LeftChar=1
+[Editor_27]
+CursorCol=1
+CursorRow=1
+TopLine=1
+LeftChar=1
+[Editor_28]
+CursorCol=21
+CursorRow=3
+TopLine=1
 LeftChar=1
