+++ inc/clue.h
@@ -1,0 +1,6 @@
+#ifndef CLUE_H_
+#define CLUE_H_
+
+
+
+#endif
