+++ inc/display.h
@@ -1,0 +1,6 @@
+#ifndef DISPLAY_H_
+#define DISPLAY_H_
+
+void display(void);
+
+#endif
