+++ inc/reaction.h
@@ -1,0 +1,6 @@
+#ifndef REACTION_H_
+#define REACTION_H_
+
+void myMouseEvent (int x, int y, int button, int event);
+
+#endif
